Move non-api classes into internal package as per convention
