+++ nicobar-groovy2/src/main/java/com/netflix/nicobar/groovy2/internal/compile/Groovy2BytecodeCompiler.java
@@ -15,7 +15,7 @@
  *     limitations under the License.
  *
  */
-package com.netflix.nicobar.groovy2.compile;
+package com.netflix.nicobar.groovy2.internal.compile;
 
 import java.io.IOException;
 import java.net.URL;
